requirements.txt
opencv-python
face_recognition
numpy
Pillow
ultralytics
tk